requirements.txt
streamlit>=1.35,<2.0
supabase>=2.4,<3.0
pandas>=2.2,<3.0
plotly>=5.20,<6.0
networkx>=3.2,<4.0
streamlit-plotly-events>=0.0.6,<1.0
pillow>=10.0,<12.0
requests>=2.31,<3.0
python-dateutil>=2.9,<3.0
